+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - `Session` class `__len__` method
 - `Session` class `__iter__` method
+### Changed
+- `tokens` parameter added to `Prompt` class
+- `tokens` parameter added to `Response` class
+- `tokens` parameter added to preset templates
 ## [0.2] - 2025-03-01
 ### Added
 - `Session` class